split up sql docs script: API + Config
