+++ sql/gen-sql-api-docs.py
@@ -16,18 +16,10 @@
 #
 
 import os
-import re
-import sys
 from collections import namedtuple
-from textwrap import dedent
-
-# To avoid adding a new direct dependency, we import markdown from within mkdocs.
-from mkdocs.structure.pages import markdown
 
 ExpressionInfo = namedtuple(
     "ExpressionInfo", "className name usage arguments examples note since deprecated")
-SQLConfEntry = namedtuple(
-    "SQLConfEntry", ["name", "default", "description"])
 
 
 def _list_function_infos(jvm):
@@ -52,18 +44,6 @@ def _list_function_infos(jvm):
             since=jinfo.getSince(),
             deprecated=jinfo.getDeprecated()))
     return sorted(infos, key=lambda i: i.name)
-
-
-def _list_sql_configs(jvm):
-    sql_configs = [
-        SQLConfEntry(
-            name=_sql_config._1(),
-            default=_sql_config._2(),
-            description=_sql_config._3(),
-        )
-        for _sql_config in jvm.org.apache.spark.sql.api.python.PythonSQLUtils.listSQLConfigs()
-    ]
-    return sql_configs
 
 
 def _make_pretty_usage(usage):
@@ -237,82 +217,11 @@ def generate_sql_markdown(jvm, path):
             mdfile.write("<br/>\n\n")
 
 
-def generate_sql_configs_table(jvm, path):
-    """
-    Generates an HTML table at `path` that lists all public SQL
-    configuration options.
-
-    The table will look something like this:
-
-    ```html
-    <table class="table">
-    <tr><th>Property Name</th><th>Default</th><th>Meaning</th></tr>
-
-    <tr>
-        <td><code>spark.sql.adaptive.enabled</code></td>
-        <td>false</td>
-        <td><p>When true, enable adaptive query execution.</p></td>
-    </tr>
-
-    ...
-
-    </table>
-    ```
-    """
-    sql_configs = _list_sql_configs(jvm)
-    value_reference_pattern = re.compile(r"^<value of (\S*)>$")
-
-    with open(path, 'w') as f:
-        f.write(dedent(
-            """
-            <table class="table">
-            <tr><th>Property Name</th><th>Default</th><th>Meaning</th></tr>
-            """
-        ))
-        for config in sorted(sql_configs, key=lambda x: x.name):
-            if config.default == "<undefined>":
-                default = "none"
-            elif config.default.startswith("<value of "):
-                referenced_config_name = value_reference_pattern.match(config.default).group(1)
-                default = "value of <code>{}</code>".format(referenced_config_name)
-            else:
-                default = config.default
-
-            if default.startswith("<"):
-                raise Exception(
-                    "Unhandled reference in SQL config docs. Config '{name}' "
-                    "has default '{default}' that looks like an HTML tag."
-                    .format(
-                        name=config.name,
-                        default=config.default,
-                    )
-                )
-
-            f.write(dedent(
-                """
-                <tr>
-                    <td><code>{name}</code></td>
-                    <td>{default}</td>
-                    <td>{description}</td>
-                </tr>
-                """
-                .format(
-                    name=config.name,
-                    default=default,
-                    description=markdown.markdown(config.description),
-                )
-            ))
-        f.write("</table>\n")
-
-
 if __name__ == "__main__":
     from pyspark.java_gateway import launch_gateway
 
     jvm = launch_gateway().jvm
-    spark_home = os.path.dirname(os.path.dirname(__file__))
+    spark_root_dir = os.path.dirname(os.path.dirname(__file__))
 
-    markdown_file_path = os.path.join(spark_home, "sql/docs/index.md")
-    sql_configs_table_path = os.path.join(spark_home, "docs/_includes/sql-configs.html")
-
+    markdown_file_path = os.path.join(spark_root_dir, "sql/docs/index.md")
     generate_sql_markdown(jvm, markdown_file_path)
-    generate_sql_configs_table(jvm, sql_configs_table_path)
